+++ sdk/storage/azure-storage-blobs/inc/azure/storage/blobs/blob_responses.hpp
@@ -82,6 +82,17 @@ namespace Azure { namespace Storage { namespace Blobs {
   public:
     Models::GetBlobPropertiesResult Value() const override { return m_pollResult; }
 
+    /**
+     * @brief Get the raw HTTP response.
+     * @return A pointer to #Azure::Core::Http::RawResponse.
+     * @note Does not give up ownership of the RawResponse.
+     */
+    Azure::Core::Http::RawResponse* GetRawResponse() const override
+    {
+      // TODO: Fix to return the rawResponse
+      return nullptr;
+    }
+
     ~StartCopyBlobOperation() override {}
 
   private: